refactor: split cron daily handler into independent job modules
Extract backup, digest, and nudge logic into lib/jobs/. Add individual
endpoints (/api/cron/backup, /api/cron/digest, /api/cron/nudges) for
manual triggering. Daily route now runs all three via Promise.allSettled
so one failure no longer aborts the others.
