+++ app/api/cron/backup/route.ts
@@ -1,0 +1,11 @@
+import { NextRequest, NextResponse } from 'next/server'
+import { checkCronAuth } from '@/lib/cron-auth'
+import { runBackupJob } from '@/lib/jobs/backup'
+
+export async function POST(request: NextRequest) {
+  const authError = checkCronAuth(request)
+  if (authError) return authError
+
+  const result = await runBackupJob()
+  return NextResponse.json({ backup: result })
+}
