+++ app/api/cron/digest/route.ts
@@ -1,0 +1,11 @@
+import { NextRequest, NextResponse } from 'next/server'
+import { checkCronAuth } from '@/lib/cron-auth'
+import { runDigestJob } from '@/lib/jobs/digest'
+
+export async function POST(request: NextRequest) {
+  const authError = checkCronAuth(request)
+  if (authError) return authError
+
+  const result = await runDigestJob()
+  return NextResponse.json({ digest: result })
+}
